chore: add menu manager

+++ la-admin/src/main/java/com/ppxb/la/admin/module/system/user/dao/UserDao.java
@@ -33,37 +33,131 @@ public interface UserDao extends BaseMapper<UserEntity> {
     void updateDisableFlag(@Param("userId") Long userId, @Param("disabledFlag") Boolean disabledFlag);
 
     /**
-     * [通过登录名查询]
+     * 通过登录名查询
      *
      * @param loginName    登录名
      * @param disabledFlag 禁用标志
-     * @return {@link UserEntity}
      * @author ppxb
      * @Date 2024/10/17 19:17
      */
     UserEntity getByLoginName(@Param("loginName") String loginName, @Param("disabledFlag") Boolean disabledFlag);
 
+    /**
+     * 通过姓名查询
+     *
+     * @param name         姓名
+     * @param disabledFlag 禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     UserEntity getByName(@Param("name") String name, @Param("disabledFlag") Boolean disabledFlag);
 
+    /**
+     * 通过手机号码查询
+     *
+     * @param phone        手机号码
+     * @param disabledFlag 禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     UserEntity getByPhone(@Param("phone") String phone, @Param("disabledFlag") Boolean disabledFlag);
 
+    /**
+     * 获取所有用户
+     *
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     List<UserVO> listAll();
 
+    /**
+     * 获取部门用户数
+     *
+     * @param departmentId 部门ID
+     * @param deletedFlag  禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     Integer countByDepartmentId(@Param("departmentId") Long departmentId, @Param("deletedFlag") Boolean deletedFlag);
 
+    /**
+     * 通过IDs获取用户
+     *
+     * @param userIds 用户IDs
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     List<UserVO> getUserByIds(@Param("userIds") Collection<Long> userIds);
 
+    /**
+     * 通过ID获取用户信息
+     *
+     * @param userId 用户ID
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     UserVO getUserById(@Param("userId") Long userId);
 
+    /**
+     * 通过部门ID获取用户
+     *
+     * @param departmentId 部门ID
+     * @param disabledFlag 禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     List<UserEntity> selectByDepartmentId(@Param("departmentId") Long departmentId, @Param("disabledFlag") Boolean disabledFlag);
 
-    List<UserEntity> selectByName(@Param("departmentIdList") List<Long> departmentIdList, @Param("name") String name, @Param("disabledFlag") Boolean disabledFlag);
+    /**
+     * 通过部门IDs和姓名查询指定用户
+     *
+     * @param departmentIdList 部门IDs
+     * @param name             姓名
+     * @param disabledFlag     禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
+    List<UserEntity> selectByName(@Param("departmentIdList") List<Long> departmentIdList,
+                                  @Param("name") String name,
+                                  @Param("disabledFlag") Boolean disabledFlag);
 
+    /**
+     * 通过部门IDs获取用户
+     *
+     * @param departmentIdList 部门IDs
+     * @param disabledFlag     禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     List<Long> getUserIdByDepartmentIdList(@Param("departmentIdList") List<Long> departmentIdList, @Param("disabledFlag") Boolean disabledFlag);
 
+    /**
+     * 获取全部用户
+     *
+     * @param leaveFlag    离职标志
+     * @param disabledFlag 禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     List<Long> getUserId(@Param("leaveFlag") Boolean leaveFlag, @Param("disabledFlag") Boolean disabledFlag);
 
+    /**
+     * 通过部门ID获取部门用户ID
+     *
+     * @param departmentId 部门ID
+     * @param disabledFlag 禁用标志
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
     List<Long> getUserIdByDepartmentId(@Param("departmentId") Long departmentId, @Param("disabledFlag") Boolean disabledFlag);
 
-    Integer updatePassword(@Param("userId") Integer userId, @Param("password") String password);
+    /**
+     * 用户重置密码
+     *
+     * @param userId   用户ID
+     * @param password 新密码
+     * @author ppxb
+     * @Date 2024/10/17 19:17
+     */
+    Integer updatePassword(@Param("userId") Long userId, @Param("password") String password);
 }
